qlib_bootstrap/requirements.txt
pyqlib==0.9.7
lightgbm==4.6.0
joblib==1.5.0
pandas==2.2.3
numpy==2.2.4
psycopg2-binary==2.9.10
PyYAML==6.0.2
minio==7.2.16
pydantic==2.11.1
pydantic-settings==2.8.1
pytest==8.3.5
